requirements.txt
pandas>=2.0.0
openpyxl>=3.1.2
Pillow>=10.0.0
requests>=2.31.0
reportlab>=4.0.0
python-dotenv>=1.0.0
